requirements/base.txt
django==2.2.12
djangorestframework<4
psycopg2-binary